backend/requirements.txt
agno
google-genai
python-dotenv
duckduckgo-search
groq
fastapi